Pack storage (#252)
* Pack EWQ ptrs

* Pack StrategyInfo into a single slot

* Pack tvl data

* Pack some rebalancing variables

* Make gov address immutable in wormhole router

+++ src/test/mocks/index.sol
@@ -17,7 +17,7 @@ contract MockL2Vault is L2Vault {
     }
 
     function setMockRebalanceDelta(uint256 _rebalanceDelta) external {
-        rebalanceDelta = _rebalanceDelta;
+        rebalanceDelta = uint224(_rebalanceDelta);
     }
 }
